% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipartite.R
\name{bipartiteRL}
\alias{bipartiteRL}
\title{Perform baseline bipartite record linkage before streaming updates}
\usage{
bipartiteRL(
  df1,
  df2,
  flds = NULL,
  flds1 = NULL,
  flds2 = NULL,
  types = NULL,
  breaks = c(0, 0.25, 0.5),
  nIter = 1000,
  burn = round(nIter * 0.1),
  a = 1,
  b = 1,
  aBM = 1,
  bBM = 1,
  seed = 0
)
}
\arguments{
\item{df1, df2}{Files 1 and 2 as dataframes where each row is a record and
each column is a field.}

\item{flds}{Names of the fields on which to compare the records in each file}

\item{flds1, flds2}{Allows specifying field names differently for each file.}

\item{types}{Types of comparisons to use for each field}

\item{breaks}{Breaks to use for Levenshtein distance on string fields}

\item{nIter, burn}{MCMC run length parameters. The returned number of samples
is nIter - burn.}

\item{a, b}{Prior parameters for m and u, respectively.}

\item{aBM, bBM}{Prior parameters for beta-linkage prior.}

\item{seed}{Random seed to set at beginning of MCMC run}
}
\description{
This function establishes a baseline linkage between two files which can be
built upon with streaming updates adding more files. It outsources the linkage
work to the BRL package and appends information to the object which will allow
streaming record linkage to continue
}
