% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{extractlinks}
\alias{extractlinks}
\title{Extract the links from a bstrlstate object into a list of streaminglinks
objects.}
\usage{
extractlinks(state)
}
\arguments{
\item{state}{A bstrlstate object returned by one of the main RL functions.}
}
\value{
A list of streaminglinks objects, one per posterior sample contained
in 'state'.
}
\description{
Extract the links from a bstrlstate object into a list of streaminglinks
objects.
}
