% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset-documentation.R
\docType{data}
\name{geco_small_result}
\alias{geco_small_result}
\title{Example linkage result object for small dataset}
\format{
An object of class 'bstrlstate', as is returned by the streaming
update functions PPRBupdate, SMCMCupdate, multifileRL, or bipartiteRL.
}
\usage{
data(geco_small_result)
}
\description{
This object is useful for examples and for testing code designed to process
record linkage results. It is identical to the PPRB result obtained from 4
files in the 'streaming-rl-howto' vignette included with this package.
}
\seealso{
geco_small, PPRBupdate
}
\keyword{datasets}
