% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPRBWTupdate.R
\name{PPRBWTupdate}
\alias{PPRBWTupdate}
\title{Perform a PPRB-with-Transition update of record linkage with a new file}
\usage{
PPRBWTupdate(
  state,
  newfile,
  flds = NULL,
  nIter.PPRB = NULL,
  burn = 0,
  blocksize = NULL,
  threestep = TRUE,
  refresh = 0.1,
  nIter.transition = 10,
  proposals.transition = c("LB", "component"),
  cores = 1,
  seed = 0,
  comparison.samples = NULL
)
}
\arguments{
\item{state}{Existing record linkage state. Returned by either bipartiteRL,
PPRBupdate, SMCMCupdate, PRPBWTupdate, or multifileRL. This is used as the
ensemble of samples for the update}

\item{newfile}{A data.frame representing the new file: one row per record}

\item{flds}{Names of fields in the new file to use for comparison. Only used
if no global field names were specified in bipartiteRL initially.}

\item{nIter.PPRB}{Number of iterations for which to run the PPRB sampler. By
default, this is the same as the number of samples present in 'state'.}

\item{burn}{Number of initial iterations to discard. The total number of
samples returned is nIter - burn.}

\item{blocksize}{Size of blocks to use for locally balanced proposals.
Default performs unblocked locally balanced proposals.}

\item{threestep}{Whether to perform three Gibbs sampling steps per iteration,
with past Z's updated with PPRB, m and u updated with full conditionals,
and the current Z updated with locally balanced proposals. If false, a two
step Gibbs sampler is used where past Z's, m and u are updated together
using PPRB and the current Z is updated with locally balanced proposals}

\item{refresh}{How often to output an update including the iteration number
and percent complete. Outputs only during the PPRB application. If
refresh >= 1, taken as a number of iterations
between messages (rounded). If 0 < refresh < 1, taken as the proportion of
nIter. If refresh == 0, no messages are displayed.}

\item{nIter.transition}{Number of iterations to use in the transition kernel
following the PPRB update for each ensemble sample.}

\item{proposals.transition}{Which kernel to use for Z
updates in the jumping and transition kernels, respectively.}

\item{cores}{Number of cores to use for parallel execution. If cores == 1,
update is run sequentially. A cluster is created using
parallel::makeCluster().}

\item{seed}{Random seed to set at the beginning of the MCMC run}

\item{comparison.samples}{A bstrlstate object for the purpose of diagnosing
the convergence of the transition kernel, relative to other samples from the
target posterior. After the PPRB step and each transition kernel
application, the KS statistic comparing each component of m to the samples
in this comparison samples object is calculated and returned in the
diagnostics.}
}
\value{
An object of class 'bstrlstate' containing posterior samples and
necessary metadata for passing to future streaming updates.
}
\description{
Perform a PPRB-with-Transition update of record linkage with a new file
}
\examples{
data(geco_small)
data(geco_small_result)

# Add fifth file to previous four-file link result
filtered <- thinsamples(geco_small_result, 2) # Filter ensemble to 2 - very small for example
file5.result <- PPRBWTupdate(filtered, geco_small[[5]],
                            nIter.PPRB=2, burn=1, nIter.transition=1, # Very small run for example
                            proposals.transition="LB",
                            blocksize=5)
}
