% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{precision}
\alias{precision}
\title{Calculate the precision of estimated links relative to true links}
\usage{
precision(sl.est, sl.true)
}
\arguments{
\item{sl.est}{streaminglinks object representing link estimates}

\item{sl.true}{streaminglinks object representing true links}
}
\value{
The precision of the estimated links.
}
\description{
Calculate the precision of estimated links relative to true links
}
