% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{randindex}
\alias{randindex}
\title{Calculate the Rand index similarity between the clusterings defined by
two sets of links}
\usage{
randindex(sl1, sl2)
}
\arguments{
\item{sl1, sl2}{Two streaming link objects defining two sets of links. E.g.
one can be a known truth and one can be an estimated link set}
}
\value{
A value between 0 and 1, the rand index of these two clusterings
}
\description{
Calculate the Rand index similarity between the clusterings defined by
two sets of links
}
\details{
Rand Index is a symmetrical measure of similarity between two
clusterings, defined as (a+b)/(n choose 2) where 'a' is the number of pairs
of elements in the same subset in both clusterings, 'b' is the number of pairs
of elements in different subsets in both clustering, and 'n' is the total
number of elements.
}
\examples{
data(geco_small)
data(geco_small_result)

sl.true <- fromentities(geco_small[[1]]$entity, geco_small[[2]]$entity,
                        geco_small[[3]]$entity, geco_small[[4]]$entity)

posterior <- extractlinks(geco_small_result)
# Compare one posterior sample to previously computed known truth
class(sl.true)
randindex(posterior[[42]], sl.true)

}
