% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocessing.R
\name{recall}
\alias{recall}
\title{Calculate the recall of estimated links relative to true links}
\usage{
recall(sl.est, sl.true)
}
\arguments{
\item{sl.est}{streaminglinks object representing link estimates}

\item{sl.true}{streaminglinks object representing true links}
}
\value{
The recall of the estimated links.
}
\description{
Calculate the recall of estimated links relative to true links
}
